Refactor checking JDA version used by libraries
